fix(chart3d): correctly interpret signed 16-bit values in GetZValue
Add explicit cast to short for proper sign extension when assembling
16-bit values from byte arrays. This ensures negative values are
correctly interpreted (e.g., 0xFFEA as -22 instead of 65514).

VAG EDC maps use signed 16-bit Big-Endian encoding for duration and
timing offsets, which requires sign extension to properly handle
negative values.

+++ Components/Chart3DComponent.cs
@@ -1067,7 +1067,9 @@ namespace VAGSuite.Components
                 // Explicitly assemble the 16-bit value from two bytes
                 int high = (int)content[index] << 8;
                 int low = (int)content[index + 1];
-                return (float)(high | low);
+                // Cast to short to ensure sign extension for negative values (e.g. 0xFFEA -> -22)
+                // Verified: VAG EDC maps use signed 16-bit Big-Endian for duration/timing offsets.
+                return (float)(short)(high | low);
             }
             else
             {
